+++ koupleless-adapter-executor-spring-2.7/src/main/java/com/alipay/sofa/koupleless/common/KouplelessThreadPoolTaskExecutor.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alipay.sofa.koupleless.common;
+
+import com.alipay.sofa.koupleless.common.util.ClassUtil;
+import com.alipay.sofa.koupleless.plugin.concurrent.KouplelessExecutorService;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.RejectedExecutionHandler;
+import java.util.concurrent.ThreadFactory;
+
+/**
+ * @author lianglipeng.llp@alibaba-inc.com
+ * @version $Id: KouplelessThreadPoolTaskExecutor.java, v 0.1 2024年05月13日 20:51 立蓬 Exp $
+ */
+public class KouplelessThreadPoolTaskExecutor extends ThreadPoolTaskExecutor {
+
+    @Override
+    protected ExecutorService initializeExecutor(ThreadFactory threadFactory,
+                                                 RejectedExecutionHandler rejectedExecutionHandler) {
+
+        ExecutorService executorService = super.initializeExecutor(threadFactory,
+            rejectedExecutionHandler);
+        KouplelessExecutorService executor = new KouplelessExecutorService(executorService);
+        ClassUtil.setField("threadPoolExecutor", this, executor);
+        return executor;
+    }
+}
