+++ koupleless-adapter-executorservice-spring-2.3-2.6/pom.xml
@@ -10,8 +10,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>koupleless-adapter-springboot-2.3</artifactId>
-    <version>${revision}</version>
+    <artifactId>koupleless-adapter-executorservice-spring-2.3-2.6</artifactId>
 
     <properties>
         <spring.boot.version>2.3.5.RELEASE</spring.boot.version>
@@ -35,7 +34,7 @@
 
         <dependency>
             <groupId>com.alipay.sofa.koupleless</groupId>
-            <artifactId>koupleless-common</artifactId>
+            <artifactId>koupleless-base-plugin</artifactId>
             <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
