--- koupleless-adapter-springboot-2.4.6/src/main/java/org/springframework/boot/autoconfigure/task/TaskSchedulingAutoConfiguration.java
@@ -1,78 +1,0 @@
-/*
- * Copyright 2012-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.boot.autoconfigure.task;
-
-import java.util.concurrent.ScheduledExecutorService;
-
-import com.alipay.sofa.koupleless.common.KouplelessThreadPoolTaskScheduler;
-import org.springframework.beans.factory.ObjectProvider;
-import org.springframework.boot.LazyInitializationExcludeFilter;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.task.TaskSchedulingProperties.Shutdown;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.task.TaskSchedulerBuilder;
-import org.springframework.boot.task.TaskSchedulerCustomizer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.annotation.SchedulingConfigurer;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.config.TaskManagementConfigUtils;
-
-/**
- * {@link EnableAutoConfiguration Auto-configuration} for {@link TaskScheduler}.
- * springboot 2.4.6 ~ 2.6.x 都可以用这个类覆盖
- * @author Stephane Nicoll
- * @since 2.1.0
- */
-@ConditionalOnClass(ThreadPoolTaskScheduler.class)
-@Configuration(proxyBeanMethods = false)
-@EnableConfigurationProperties(TaskSchedulingProperties.class)
-@AutoConfigureAfter(TaskExecutionAutoConfiguration.class)
-public class TaskSchedulingAutoConfiguration {
-
-    @Bean
-    @ConditionalOnBean(name = TaskManagementConfigUtils.SCHEDULED_ANNOTATION_PROCESSOR_BEAN_NAME)
-    @ConditionalOnMissingBean({ SchedulingConfigurer.class, TaskScheduler.class, ScheduledExecutorService.class })
-    public ThreadPoolTaskScheduler taskScheduler(TaskSchedulerBuilder builder) {
-        return builder.configure(new KouplelessThreadPoolTaskScheduler());
-    }
-
-    @Bean
-    public static LazyInitializationExcludeFilter scheduledBeanLazyInitializationExcludeFilter() {
-        return new ScheduledBeanLazyInitializationExcludeFilter();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    public TaskSchedulerBuilder taskSchedulerBuilder(TaskSchedulingProperties properties,
-                                                     ObjectProvider<TaskSchedulerCustomizer> taskSchedulerCustomizers) {
-        TaskSchedulerBuilder builder = new TaskSchedulerBuilder();
-        builder = builder.poolSize(properties.getPool().getSize());
-        Shutdown shutdown = properties.getShutdown();
-        builder = builder.awaitTermination(shutdown.isAwaitTermination());
-        builder = builder.awaitTerminationPeriod(shutdown.getAwaitTerminationPeriod());
-        builder = builder.threadNamePrefix(properties.getThreadNamePrefix());
-        builder = builder.customizers(taskSchedulerCustomizers);
-        return builder;
-    }
-
-}
